requirements.txt
# Core Python and Notebook Utilities
ipykernel
numpy
scipy
matplotlib
seaborn
plotly

# Machine Learning / Modeling
scikit-learn
imblearn
xgboost
catboost

# Natural Language Processing and Embeddings
spacy
transformers
sentence-transformers
faiss-cpu

# Generative AI and API integration
openai

# Data storage and cloud
pymongo
boto3
mypy-boto3-s3
botocore

# Monitoring and MLOps
mlflow
neuro_mf

# Utilities and Config Management
dill
PyYAML
from_root
loguru

# Backend / Web Application
fastapi
uvicorn
jinja2
python-multipart
gradio

# Visualization and Debugging
matplotlib
plotly
seaborn

# --- Core Framework ---
langchain>=1.1.3
langchain-core>=1.1.3
langchain-community>=0.4.1
langchain-text-splitters>=1.0.0

# --- Vector Database & Model Integrations ---
langchain-pinecone>=0.2.8
langchain-google-genai>=3.1.0
langchain-openai>=1.1.1
pinecone-client>=5.0.0

# --- Embeddings & NLP ---
sentence-transformers>=3.0.0
huggingface-hub>=0.26.0
nltk>=3.9.0
pandas>=2.2.0

# --- Web & Utilities ---
flask==3.1.1
flask-cors>=5.0.0
python-dotenv>=1.0.1
pypdf>=5.0.0
requests>=2.32.0


# Local Package Installation
-e .    